Structural and reorganized update
Changed colors to BaseLayout as it is more appropriate to its purpose. Moved every "Base" layout definition there and changed the names of several ids and classes to match their semantic role in the html.

Also took out the main title to use it as a name holder in the desktop version.