expandable CVs, individual page logic and template
Made the show button to work in the index page to expand the pdfs of my (not updated) CVs

I also decided that this logic was not pertinent to a "general page" on my website and therefore decided to made it specific for the index.

Same idea applies to the style (css) so I decided to make a template.html from which I should be able to easily start my new pages on the site.